requirements.txt
mcstatus
py-cord
python-aternos
python-dotenv
